resolve #10 ユーザーをDBから取得できるようにする

+++ src/main/kotlin/com/ryohysk/graphqlkotlin/domain/repository/PointRepository.kt
@@ -1,5 +1,5 @@
 package com.ryohysk.graphqlkotlin.domain.repository
 
 interface PointRepository {
-    fun findUsersPointsMap(): Map<Int, Int>
+    fun findUsersPointsMap(): Map<Long, Int>
 }
